Limit requests for ATXs globally
